requirements.txt
django==1.8
djangorestframework==3.1
djangorestframework-jwt==1.4
dj-database-url==0.3
path.py==7.2
psycopg2==2.5
pyjwt==0.4
requests==2.6
uwsgi==2.0
